Revert "Fusion PouvoirFigure et PouvoirFigureMini en un seul composant"
This reverts commit 30cf2c0c8392a9903e5828f0aea42187f8ec21b7.
